graspevaluator/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
textstat>=0.7.3
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
argparse>=1.4.0
datetime
pathlib2
json5>=0.9.0
plotly>=5.17.0